打包第三方库 支持npm打包
