+++ example/index.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <title>TypeScript + Webpack 4</title>
+    <style>
+        html, body {margin:0; padding: 0; height: 100%}
+    </style>
+</head>
+
+<body>
+    <canvas id="glcanvas"  ></canvas>
+</body>
+
+</html>
