+++ package.json
@@ -1,11 +1,13 @@
 {
-  "name": "xmapeditorgo",
-  "version": "1.0.0",
+  "name": "@megvii/glview",
+  "version": "1.0.5",
   "description": "",
-  "main": "index.ts",
+  "main": "dist/main.js",
+  "module": "dist/main.js",
+  "types": "dist/types/src/index.d.ts",
   "scripts": {
-    "build": "webpack",
-    "dev": "webpack-dev-server",
+    "dev": "NODE_ENV=development webpack-dev-server --config ./webpack.config.js",
+    "build": "NODE_ENV=production webpack --config ./webpack.build.config.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
@@ -18,5 +20,7 @@
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.10"
   },
-  "dependencies": {}
+  "dependencies": {
+    "html-webpack-plugin": "^3.2.0"
+  }
 }
