+++ webpack.build.config.js
@@ -1,0 +1,25 @@
+const path = require('path');
+const webpack = require('webpack');
+
+module.exports = {
+    mode: 'development',
+
+    entry: './src/index.ts',
+    output: {
+        filename: 'main.js',
+        libraryTarget: 'umd',
+        path: path.resolve(__dirname, 'dist')
+    },
+
+    module: {
+        rules: [{
+            test: /\.ts$/,
+            use: "ts-loader"
+        }]
+    },
+    resolve: {
+        extensions: [
+            '.ts', '.js', '.html',
+        ]
+    },
+};
